+++ src/components/ButtonBase/ButtonBase.tsx
@@ -1,9 +1,16 @@
 import React from 'react';
+import { SemanticColorToken } from 'src/utils/new-colors';
 import { GradeBand } from 'src/enum/gradeband';
 
 export type ButtonBaseProps = {
   children?: React.ReactNode;
+  id?: string;
   classes?: string;
+  variant?: 'text' | 'contained' | 'outlined' | 'plain';
+  state?: 'activated' | 'visited';
+  color?: SemanticColorToken;
+  disableElevation?: boolean;
+  size?: 'sm' | 'md';
   disabled?: boolean;
   title?: string;
   onClick?: (event: React.MouseEvent<HTMLButtonElement>) => void;
@@ -23,14 +30,20 @@ export type ButtonBaseProps = {
   ariaExpanded?: boolean;
   ariaControls?: string;
   ariaPressed?: boolean;
-  gradeBand?: GradeBand;
   role?: string;
   autoFocus?: boolean;
+  gradeBand?: GradeBand;
 };
 
 export const ButtonBase: React.FC<ButtonBaseProps> = ({
   children,
+  id,
   classes,
+  variant,
+  state,
+  color,
+  size = 'md',
+  disableElevation = false,
   disabled = false,
   title,
   onClick,
@@ -54,9 +67,22 @@ export const ButtonBase: React.FC<ButtonBaseProps> = ({
   autoFocus,
   ...props
 }) => {
+  const buttonClasses = [
+    'connect__button',
+    variant && `connect__button-${variant}`,
+    state && `connect__button-${state}`,
+    color && `connect__button-${color}`,
+    disabled && 'connect__button-disabled',
+    size === 'sm' && 'connect__button-small',
+    disableElevation && 'connect__button-no-elevation',
+  ]
+    .filter(Boolean)
+    .join(' ');
+
   return (
     <Component
-      className={classes}
+      id={id}
+      className={`${buttonClasses} ${classes}`}
       disabled={Component === 'button' ? disabled : undefined}
       aria-disabled={disabled}
       onClick={disabled ? undefined : onClick}